# [chore][exporter/splunkhec] reuse metric buffer